requirements.txt
# Telegram Bot
python-telegram-bot==20.7
telegram==0.0.1

# Data APIs
yfinance==0.2.28
requests==2.31.0
beautifulsoup4==4.12.2

# Crypto APIs
pycoingecko==3.1.0
python-binance==1.0.19
ccxt==4.1.77

# AI/ML
openai>=1.35.0
anthropic==0.30.1

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Database and Caching
redis==5.0.1
sqlite3

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Scheduling and Background Tasks
APScheduler==3.10.4
pytz==2023.3

# Logging and Monitoring
loguru==0.7.2

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Time and Date
python-dateutil==2.8.2

# Configuration
configparser==6.0.0 